Move NotificationManager
